Lógica - Save & Load
Implementação das funções de Save e Load Game. O Save é disparado pela letra "T" e o Load pela "Ç", falta colocar as chamadas nos pontos corretos e no begin play, tentei colocar o Load no begin play, mas está dando Crash no editor.

+++ Source/CoopGame/Private/SPlayerState.cpp
@@ -16,12 +16,12 @@ void ASPlayerState::SubScore(float ScoreDelta)
 }
 
 
-void ASPlayerState::AddCredits(int32 CreditsDelta)
+void ASPlayerState::AddCredits(float CreditsDelta)
 {
 	PLayerCredits = PLayerCredits + CreditsDelta;
 }
 
-void ASPlayerState::SubCredits(int32 CreditsDelta)
+void ASPlayerState::SubCredits(float CreditsDelta)
 {
 	PLayerCredits = PLayerCredits - CreditsDelta;
 }
